OpenUSD Autoupgrade: Let it rip & begin damage control.

+++ Sources/Arch/systemInfo.cpp
@@ -1,28 +1,11 @@
 //
 // Copyright 2016 Pixar
 //
-// Licensed under the Apache License, Version 2.0 (the "Apache License")
-// with the following modification; you may not use this file except in
-// compliance with the Apache License and the following modification to it:
-// Section 6. Trademarks. is deleted and replaced with:
-//
-// 6. Trademarks. This License does not grant permission to use the trade
-//    names, trademarks, service marks, or product names of the Licensor
-//    and its affiliates, except as required to comply with Section 4(c) of
-//    the License and to reproduce the content of the NOTICE file.
-//
-// You may obtain a copy of the Apache License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the Apache License with the above modification is
-// distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-// KIND, either express or implied. See the Apache License for the specific
-// language governing permissions and limitations under the Apache License.
+// Licensed under the terms set forth in the LICENSE.txt file available at
+// https://openusd.org/license.
 //
 
-#include <pxr/pxrns.h>
+#include "pxr/pxrns.h"
 
 #include "Arch/error.h"
 #include "Arch/fileSystem.h"
@@ -144,7 +127,7 @@ std::string ArchGetExecutablePath()
 
   // On Darwin _NSGetExecutablePath() returns the executable path.
   return _DynamicSizedRead(ARCH_PATH_MAX, [](char *buffer, size_t *size) {
-    uint32_t bufsize = static_cast<unsigned int>(*size);
+    uint32_t bufsize = *size;
     if (_NSGetExecutablePath(buffer, &bufsize) == -1) {
       // We're told the correct size.
       *size = bufsize;
@@ -184,7 +167,7 @@ std::string ArchGetExecutablePath()
 int ArchGetPageSize()
 {
 #if defined(ARCH_OS_LINUX) || defined(ARCH_OS_DARWIN)
-  return static_cast<int>(sysconf(_SC_PAGE_SIZE));
+  return sysconf(_SC_PAGE_SIZE);
 #elif defined(ARCH_OS_WINDOWS)
   SYSTEM_INFO info;
   GetSystemInfo(&info);
