+++ Sources/Trace/aggregateNode.cpp
@@ -1,30 +1,13 @@
 //
 // Copyright 2018 Pixar
 //
-// Licensed under the Apache License, Version 2.0 (the "Apache License")
-// with the following modification; you may not use this file except in
-// compliance with the Apache License and the following modification to it:
-// Section 6. Trademarks. is deleted and replaced with:
+// Licensed under the terms set forth in the LICENSE.txt file available at
+// https://openusd.org/license.
 //
-// 6. Trademarks. This License does not grant permission to use the trade
-//    names, trademarks, service marks, or product names of the Licensor
-//    and its affiliates, except as required to comply with Section 4(c) of
-//    the License and to reproduce the content of the NOTICE file.
-//
-// You may obtain a copy of the Apache License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the Apache License with the above modification is
-// distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-// KIND, either express or implied. See the Apache License for the specific
-// language governing permissions and limitations under the Apache License.
-//
-
-#include <pxr/pxrns.h>
 
 #include "Trace/aggregateNode.h"
+
+#include "pxr/pxrns.h"
 
 #include <vector>
 
@@ -185,7 +168,7 @@ void TraceAggregateNode::AdjustForOverheadAndNoise(TimeStamp scopeOverhead,
 struct _StackNode {
   _StackNode(TraceAggregateNodePtr node, int parentIdx) : eventNode(node), parentIdx(parentIdx)
   {
-    remainingChildren = static_cast<int>(node->GetChildrenRef().size());
+    remainingChildren = node->GetChildrenRef().size();
   }
 
   TraceAggregateNodePtr eventNode;
@@ -262,7 +245,7 @@ void TraceAggregateNode::MarkRecursiveChildren()
       // Here our node has children, so before we go on, we must
       // push our children on the child stack.  This gives us the
       // post-order traversal we need.
-      int parent = static_cast<int>(stack.size()) - 1;
+      int parent = stack.size() - 1;
       for (int i = 0; i < numKids; i++) {
         // Only process nodes that have not been previously processed
         // (by a previous call to Report() for example).  If a node
